requirements.txt
requests>=2.28.0
feedparser>=6.0.0
beautifulsoup4>=4.11.0
python-dotenv>=0.20.0
premailer>=3.10.0
apscheduler>=3.10.0
